remove tsx, making generateSchema work
