+++ packages/grafana-data/scripts/generateSchema.mts
@@ -2,7 +2,7 @@ import fs from 'node:fs';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 
-import { NewThemeOptionsSchema } from '../src/themes/createTheme';
+import { NewThemeOptionsSchema } from '../src/themes/types/schema.mts';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
