+++ packages/grafana-data/src/themes/breakpoints.ts
@@ -1,28 +1,4 @@
-/** @beta */
-export interface ThemeBreakpointValues {
-  xs: number;
-  sm: number;
-  md: number;
-  lg: number;
-  xl: number;
-  xxl: number;
-}
-
-/** @beta */
-export type ThemeBreakpointsKey = keyof ThemeBreakpointValues;
-
-/** @beta */
-export interface ThemeBreakpoints {
-  values: ThemeBreakpointValues;
-  keys: string[];
-  unit: string;
-  up: (key: ThemeBreakpointsKey | number) => string;
-  down: (key: ThemeBreakpointsKey | number) => string;
-  container: {
-    up: (key: ThemeBreakpointsKey | number, name?: string) => string;
-    down: (key: ThemeBreakpointsKey | number, name?: string) => string;
-  };
-}
+import { type ThemeBreakpoints, type ThemeBreakpointsKey, type ThemeBreakpointValues } from './types/breakpoints.mts';
 
 /** @internal */
 export function createBreakpoints(): ThemeBreakpoints {
