+++ packages/grafana-data/src/themes/createColors.ts
@@ -1,119 +1,8 @@
 import { merge } from 'lodash';
-import { z } from 'zod';
 
 import { alpha, darken, emphasize, getContrastRatio, lighten } from './colorManipulator';
 import { palette } from './palette';
-import { type DeepRequired, type ThemeRichColor, ThemeRichColorInputSchema } from './types';
-
-const ThemeColorsModeSchema = z.enum(['light', 'dark']);
-/** @internal */
-type ThemeColorsMode = z.infer<typeof ThemeColorsModeSchema>;
-
-const createThemeColorsBaseSchema = <TColor>(color: TColor) =>
-  z
-    .object({
-      mode: ThemeColorsModeSchema,
-
-      primary: color,
-      secondary: color,
-      tertiary: color,
-      info: color,
-      error: color,
-      success: color,
-      warning: color,
-
-      text: z.object({
-        primary: z.string().optional(),
-        secondary: z.string().optional(),
-        disabled: z.string().optional(),
-        link: z.string().optional(),
-        /** Used for auto white or dark text on colored backgrounds */
-        maxContrast: z.string().optional(),
-      }),
-
-      background: z.object({
-        /** Dashboard and body background */
-        canvas: z.string().optional(),
-        /** Primary content pane background (panels etc) */
-        primary: z.string().optional(),
-        /** Cards and elements that need to stand out on the primary background */
-        secondary: z.string().optional(),
-        /**
-         * For popovers and menu backgrounds. This is the same color as primary in most light themes but in dark
-         * themes it has a brighter shade to help give it contrast against the primary background.
-         **/
-        elevated: z.string().optional(),
-      }),
-
-      border: z.object({
-        weak: z.string().optional(),
-        medium: z.string().optional(),
-        strong: z.string().optional(),
-      }),
-
-      gradients: z.object({
-        brandVertical: z.string().optional(),
-        brandHorizontal: z.string().optional(),
-      }),
-
-      action: z.object({
-        /** Used for selected menu item / select option */
-        selected: z.string().optional(),
-        /**
-         * @alpha (Do not use from plugins)
-         * Used for selected items when background only change is not enough (Currently only used for FilterPill)
-         **/
-        selectedBorder: z.string().optional(),
-        /** Used for hovered menu item / select option */
-        hover: z.string().optional(),
-        /** Used for button/colored background hover opacity */
-        hoverOpacity: z.number().optional(),
-        /** Used focused menu item / select option */
-        focus: z.string().optional(),
-        /** Used for disabled buttons and inputs */
-        disabledBackground: z.string().optional(),
-        /** Disabled text */
-        disabledText: z.string().optional(),
-        /** Disablerd opacity */
-        disabledOpacity: z.number().optional(),
-      }),
-
-      scrollbar: z.string().optional(),
-      hoverFactor: z.number(),
-      contrastThreshold: z.number(),
-      tonalOffset: z.number(),
-    })
-    .partial();
-
-// Need to override the zod type to include the generic properly
-/** @internal */
-type ThemeColorsBase<TColor> = DeepRequired<
-  Omit<
-    z.infer<ReturnType<typeof createThemeColorsBaseSchema>>,
-    'primary' | 'secondary' | 'tertiary' | 'info' | 'error' | 'success' | 'warning'
-  >
-> & {
-  primary: TColor;
-  secondary: TColor;
-  tertiary: TColor;
-  info: TColor;
-  error: TColor;
-  success: TColor;
-  warning: TColor;
-};
-
-/** @beta */
-export interface ThemeColors extends ThemeColorsBase<ThemeRichColor> {
-  /** Returns a text color for the background */
-  getContrastText(background: string, threshold?: number): string;
-  /* Brighten or darken a color by specified factor (0-1) */
-  emphasize(color: string, amount?: number): string;
-}
-
-export const ThemeColorsInputSchema = createThemeColorsBaseSchema(ThemeRichColorInputSchema);
-
-/** @internal */
-export type ThemeColorsInput = z.infer<typeof ThemeColorsInputSchema>;
+import { type ThemeColorsMode, type ThemeRichColor, type ThemeColorsBase } from './types/color.mts';
 
 class DarkColors implements ThemeColorsBase<Partial<ThemeRichColor>> {
   mode: ThemeColorsMode = 'dark';
