+++ packages/grafana-data/src/themes/createComponents.ts
@@ -1,67 +1,6 @@
-import { type ThemeColors } from './createColors';
 import { type ThemeShadows } from './createShadows';
-import type { Radii } from './createShape';
-import type { ThemeSpacingTokens } from './createSpacing';
-
-interface MenuComponentTokens {
-  borderRadius: keyof Radii;
-  padding: ThemeSpacingTokens;
-}
-
-/** @beta */
-export interface ThemeComponents {
-  /** Applies to normal buttons, inputs, radio buttons, etc */
-  height: {
-    sm: number;
-    md: number;
-    lg: number;
-  };
-  input: {
-    background: string;
-    borderColor: string;
-    borderHover: string;
-    text: string;
-  };
-  tooltip: {
-    text: string;
-    background: string;
-  };
-  panel: {
-    padding: number;
-    headerHeight: number;
-    borderColor: string;
-    boxShadow: string;
-    background: string;
-  };
-  dropdown: {
-    background: string;
-  };
-  overlay: {
-    background: string;
-  };
-  dashboard: {
-    background: string;
-    padding: number;
-  };
-  drawer: {
-    padding: number;
-  };
-  textHighlight: {
-    background: string;
-    text: string;
-  };
-  sidemenu: {
-    width: number;
-  };
-  horizontalDrawer: {
-    defaultHeight: number;
-  };
-  table: {
-    rowHoverBackground: string;
-    rowSelected: string;
-  };
-  menu: MenuComponentTokens;
-}
+import { type ThemeColors } from './types/color.mts';
+import { type MenuComponentTokens, type ThemeComponents } from './types/components.mts';
 
 export function createComponents(colors: ThemeColors, shadows: ThemeShadows): ThemeComponents {
   const panel = {
