+++ packages/grafana-data/src/themes/createTypography.ts
@@ -1,60 +1,9 @@
 // Code based on Material UI
 // The MIT License (MIT)
 // Copyright (c) 2014 Call-Em-All
-import { z } from 'zod';
-
-import { type ThemeColors } from './createColors';
-
-/** @beta */
-export interface ThemeTypography extends ThemeTypographyVariantTypes {
-  fontFamily: string;
-  fontFamilyMonospace: string;
-  fontSize: number;
-  fontWeightLight: number;
-  fontWeightRegular: number;
-  fontWeightMedium: number;
-  fontWeightBold: number;
-
-  // The font-size on the html element.
-  htmlFontSize?: number;
-
-  /**
-   * @deprecated
-   * from legacy old theme
-   * */
-  size: {
-    base: string;
-    xs: string;
-    sm: string;
-    md: string;
-    lg: string;
-  };
-
-  pxToRem: (px: number) => string;
-}
-
-export interface ThemeTypographyVariant {
-  fontSize: string;
-  fontWeight: number;
-  lineHeight: number;
-  fontFamily: string;
-  letterSpacing?: string;
-}
-
-export const ThemeTypographyInputSchema = z.object({
-  fontFamily: z.string().optional(),
-  fontFamilyMonospace: z.string().optional(),
-  fontSize: z.number().positive().optional(),
-  fontWeightLight: z.number().positive().optional(),
-  fontWeightRegular: z.number().positive().optional(),
-  fontWeightMedium: z.number().positive().optional(),
-  fontWeightBold: z.number().positive().optional(),
-  // what's the font-size on the html element.
-  // 16px is the default font-size used by browsers.
-  htmlFontSize: z.number().positive().optional(),
-});
-
-export type ThemeTypographyInput = z.infer<typeof ThemeTypographyInputSchema>;
+import { type ThemeColors } from './types/color.mts';
+import { type ThemeTypographyInput } from './types/schema.mts';
+import { type ThemeTypography, type ThemeTypographyVariant } from './types/typography.mts';
 
 const defaultFontFamily = "'Inter', 'Helvetica', 'Arial', sans-serif";
 const defaultFontFamilyMonospace = "'Roboto Mono', monospace";
@@ -145,16 +94,4 @@ export function createTypography(colors: ThemeColors, typographyInput: ThemeTypo
 
 function round(value: number) {
   return Math.round(value * 1e5) / 1e5;
-}
-
-export interface ThemeTypographyVariantTypes {
-  h1: ThemeTypographyVariant;
-  h2: ThemeTypographyVariant;
-  h3: ThemeTypographyVariant;
-  h4: ThemeTypographyVariant;
-  h5: ThemeTypographyVariant;
-  h6: ThemeTypographyVariant;
-  body: ThemeTypographyVariant;
-  bodySmall: ThemeTypographyVariant;
-  code: ThemeTypographyVariant;
 }
