+++ packages/grafana-data/src/themes/types/shadows.mts
@@ -1,0 +1,6 @@
+/** @beta */
+export interface ThemeShadows {
+  z1: string;
+  z2: string;
+  z3: string;
+}
